R6 class for a dbi connection, see connector_dbi (used to interact with DBI compliant database backends)

Connector object for DBI connections. This object is used to interact with DBI compliant database backends. See the DBI package for how which backends are supported.

Details

Upon garbage collection, the connection will try to disconnect from the database. But it is good practice to call disconnect when you are done with the connection.

Methods

Public methods

Connector_dbi$new()

Connector_dbi$list_content()

Connector_dbi$get_conn()

Connector_dbi$disconnect()

Connector_dbi$read()

Connector_dbi$write()

Connector_dbi$remove()

Connector_dbi$tbl()

Method new()

Initialize the connection

Usage

Connector_dbi$new(drv, ...)

Arguments

drv

DBI driver

...

Additional arguments passed to DBI::dbConnect

Returns

A connector_dbi object

Method list_content()

List tables in the database

Usage

Connector_dbi$list_content(...)

Arguments

...

Additional arguments passed to DBI::dbListTables

Returns

A character vector of table names

Method get_conn()

Get the connection object

Usage

Connector_dbi$get_conn()

Returns

A DBI connection object

Method disconnect()

Disconnect from the database

Usage

Connector_dbi$disconnect()

Method read()

Read a table from the database

Usage

Connector_dbi$read(name, ...)

Arguments

name

character Table name

...

Additional arguments passed to DBI::dbReadTable

Returns

A data.frame

Method write()

Write a table to the database

Usage

Connector_dbi$write(x, name, ...)

Arguments

x

data.frame Table to write

name

character Table name

...

Additional arguments passed to DBI::dbWriteTable

Method remove()

Remove a table from the database

Usage

Connector_dbi$remove(name, ...)

Arguments

name

character Table name

...

Additional arguments passed to DBI::dbRemoveTable

Method tbl()

Create a tbl object

Usage

Connector_dbi$tbl(name, ...)

Arguments

name

character Table name

...

Additional arguments passed to dplyr::tbl

Examples

# Create DBI connector

db <- connector::Connector_dbi$new(RSQLite::SQLite(), ":memory:")

db
#> <Connector_dbi>
#>   Public:
#>     disconnect: function () 
#>     get_conn: function () 
#>     initialize: function (drv, ...) 
#>     list_content: function (...) 
#>     read: function (name, ...) 
#>     remove: function (name, ...) 
#>     tbl: function (name, ...) 
#>     write: function (x, name, ...) 
#>   Private:
#>     conn: SQLiteConnection
#>     finalize: function () 

# Write to the database

db$write(iris, "iris")

# Read from the database

db$read("iris") |>
  head(5)
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa

# List available tables

db$list_content()
#> [1] "iris"

# Use the connector to run a query

db$get_conn() |>
  DBI::dbGetQuery("SELECT * FROM iris limit 5")
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa

# Use dplyr verbs and collect data

db$tbl("iris") |>
  dplyr::filter(Sepal.Length > 7) |>
  dplyr::collect()
#> # A tibble: 12 × 5
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width Species  
#>           <dbl>       <dbl>        <dbl>       <dbl> <chr>    
#>  1          7.1         3            5.9         2.1 virginica
#>  2          7.6         3            6.6         2.1 virginica
#>  3          7.3         2.9          6.3         1.8 virginica
#>  4          7.2         3.6          6.1         2.5 virginica
#>  5          7.7         3.8          6.7         2.2 virginica
#>  6          7.7         2.6          6.9         2.3 virginica
#>  7          7.7         2.8          6.7         2   virginica
#>  8          7.2         3.2          6           1.8 virginica
#>  9          7.2         3            5.8         1.6 virginica
#> 10          7.4         2.8          6.1         1.9 virginica
#> 11          7.9         3.8          6.4         2   virginica
#> 12          7.7         3            6.1         2.3 virginica

# Disconnect from the database

db$disconnect()