Class Connector_fs

The connector_fs class is a file system connector for accessing and manipulating files in a local file system.

Methods

Public methods

Connector_fs$new()

Connector_fs$list_content()

Connector_fs$construct_path()

Connector_fs$read()

Connector_fs$write()

Connector_fs$remove()

Connector_fs$get_path()

Connector_fs$clone()

Method new()

Initializes the connector_fs class

Usage

Connector_fs$new(path, access = "rw")

Arguments

path

Path to the file system

access

Access type ("rw" by default)

Method list_content()

Returns the list of files in the specified path

Usage

Connector_fs$list_content(...)

Arguments

...

Other parameters to pass to the list.files function

Method construct_path()

Constructs a complete path by combining the specified access path with the provided elements

Usage

Connector_fs$construct_path(...)

Arguments

...

Elements to construct the path

Method read()

Reads the content of the specified file using the private access path and additional options

Usage

Connector_fs$read(name, ...)

Arguments

name

Name of the file to read

...

Other parameters to pass to the read_file function (depends on the extension of a file)

Method write()

Writes the specified content to the specified file using the private access path and additional options

Usage

Connector_fs$write(x, file, ...)

Arguments

x

Content to write to the file

file

File name

...

Other parameters to pass to the write_file function (depends on the extension of a file)

Method remove()

Remove the specified file by given name with extension

Usage

Connector_fs$remove(file)

Arguments

file

File name

Method get_path()

Returns the path of the file system

Usage

Connector_fs$get_path()

Returns

Path to the file system

Method clone()

The objects of this class are cloneable with this method.

Usage

Connector_fs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.