Read content from the connector

Generic implementing of how to read content from the different connector objects:

connector_dbi: Uses DBI::dbReadTable() to read the table from the DBI connection.

connector_fs: Uses read_file() to read a given file. The underlying function used, and thereby also the arguments available through ... depends on the file extension.

Usage

cnt_read(connector_object, name, ...)

# S3 method for class 'connector_dbi'
cnt_read(connector_object, name, ...)

# S3 method for class 'connector_fs'
cnt_read(connector_object, name, ...)

Arguments

connector_object

connector The connector object to use.

name

character Name of the content to read, write, or remove. Typically the table name.

...

Additional arguments passed to the method for the individual connector.

Value

R object with the content. For rectangular data a data.frame.

Examples

# Read table from DBI database
cnt <- connector_dbi$new(RSQLite::SQLite())

cnt |>
  cnt_write(iris, "iris")

cnt |>
  cnt_list_content()
#> [1] "iris"

cnt |>
  cnt_read("iris") |>
  head()
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa
#> 6          5.4         3.9          1.7         0.4  setosa

# Write and read a CSV file using the file storage connector
cnt <- connector_fs$new(tempdir())

cnt |>
  cnt_write(iris, "iris.csv")

cnt |>
  cnt_read("iris.csv") |>
  head()
#> Rows: 150 Columns: 5
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (1): Species
#> dbl (4): Sepal.Length, Sepal.Width, Petal.Length, Petal.Width
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 6 × 5
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#>          <dbl>       <dbl>        <dbl>       <dbl> <chr>  
#> 1          5.1         3.5          1.4         0.2 setosa 
#> 2          4.9         3            1.4         0.2 setosa 
#> 3          4.7         3.2          1.3         0.2 setosa 
#> 4          4.6         3.1          1.5         0.2 setosa 
#> 5          5           3.6          1.4         0.2 setosa 
#> 6          5.4         3.9          1.7         0.4 setosa