General connector object

This R6 class is a general class for all connectors. It is used to define the methods that all connectors should have. New connectors should inherit from this class, and the methods described in connector_methods should be implemented.

See also

Methods

Public methods

connector$new()

connector$cnt_list_content()

connector$cnt_read()

connector$cnt_write()

connector$cnt_remove()

connector$clone()

Method new()

Initialize connector with the option of adding an extra class

Usage

connector$new(extra_class = NULL)

Arguments

extra_class

character Extra class to assign to the new connector.

Method cnt_list_content()

List content

Usage

connector$cnt_list_content(...)

Arguments

...

Additional arguments passed to the method for the individual connector.

Returns

A character vector of content names

Method cnt_read()

Read content

Usage

connector$cnt_read(name, ...)

Arguments

name

character Name of the content to read, write, or remove. Typically the table name.

...

Additional arguments passed to the method for the individual connector.

Returns

The result of the read method

Method cnt_write()

Write content

Usage

connector$cnt_write(x, name, ...)

Arguments

x

The object to write to the connection

name

character Name of the content to read, write, or remove. Typically the table name.

...

Additional arguments passed to the method for the individual connector.

Method cnt_remove()

Remove or delete content

Usage

connector$cnt_remove(name, ...)

Arguments

name

character Name of the content to read, write, or remove. Typically the table name.

...

Additional arguments passed to the method for the individual connector.

Method clone()

The objects of this class are cloneable with this method.

Usage

connector$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.