Read content from the connector

Generic implementing of how to read content from the different connector objects:

ConnectorDBI: Uses DBI::dbReadTable() to read the table from the DBI connection.

ConnectorFS: Uses read_file() to read a given file. The underlying function used, and thereby also the arguments available through ... depends on the file extension.

ConnectorLogger: Logs the read operation and calls the underlying connector method.

Usage

read_cnt(connector_object, name, ...)

# S3 method for class 'ConnectorDBI'
read_cnt(connector_object, name, ...)

# S3 method for class 'ConnectorFS'
read_cnt(connector_object, name, ...)

# S3 method for class 'ConnectorLogger'
read_cnt(connector_object, name, ...)

Arguments

connector_object

Connector The connector object to use.

name

character Name of the content to read, write, or remove. Typically the table name.

...

Additional arguments passed to the method for the individual connector.

Value

R object with the content. For rectangular data a data.frame.

Examples

# Read table from DBI database
cnt <- connector_dbi(RSQLite::SQLite())

cnt |>
  write_cnt(iris, "iris")

cnt |>
  list_content_cnt()
#> [1] "iris"

cnt |>
  read_cnt("iris") |>
  head()
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa
#> 6          5.4         3.9          1.7         0.4  setosa

# Write and read a CSV file using the file storage connector

folder <- withr::local_tempdir()
cnt <- connector_fs(folder)
#> Error in validate_resource(self): Invalid file system connector: /tmp/RtmpNYsw8N/file1e9a659823d2 does not
#> exist.

cnt |>
  write_cnt(iris, "iris.csv")

cnt |>
  read_cnt("iris.csv") |>
  head()
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa
#> 6          5.4         3.9          1.7         0.4  setosa

# Add logging to a file system connector
folder <- withr::local_tempdir()
cnt <- connectors(data = connector_fs(folder)) |> add_logs()
#> Error in validate_resource(self): Invalid file system connector: /tmp/RtmpNYsw8N/file1e9a187964b7 does not
#> exist.

cnt$data |>
  write_cnt(iris, "iris.csv")
#> Error in method_error_msg(connector_object): Method not implemented for class <NULL>
#> ℹ See the customize (`vignette(connector::customize)`) vignette on how to
#>   create custom connectors and methods

cnt$data |>
  read_cnt("iris.csv") |>
  head()
#> Error in method_error_msg(connector_object): Method not implemented for class <NULL>
#> ℹ See the customize (`vignette(connector::customize)`) vignette on how to
#>   create custom connectors and methods