Create FS connector

Create a new FS connector object. See Connector_fs for details.

Usage

connector_fs(path, ..., extra_class = NULL)

Arguments

path

Path to the file system

...

Additional arguments passed to Connector_fs

extra_class

character Extra class added to the object. See details.

Value

A new Connector_fs object

Details

The extra_class parameter allows you to create a subclass of the Connector_fs object. This can be useful if you want to create a custom connection object for easier dispatch of new s3 methods, while still inheriting the methods from the Connector_fs object.

Examples

# Connect to a file system

path_to_adam <- system.file("demo_trial", "adam", package = "connector")
db <- connector_fs(path_to_adam)

db
#> <Connector_fs>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     construct_path: function (...) 
#>     get_path: function () 
#>     initialize: function (path, access = "rw") 
#>     list_content: function (...) 
#>     read: function (name, ...) 
#>     remove: function (file) 
#>     write: function (x, file, ...) 
#>   Private:
#>     path: /home/runner/work/_temp/Library/connector/demo_trial/adam

# Create subclass connection

db_subclass <- connector_fs(path_to_adam, extra_class = "subclass")

db_subclass
#> <subclass>
#>   Inherits from: <Connector_fs>
#>   Public:
#>     clone: function (deep = FALSE) 
#>     construct_path: function (...) 
#>     get_path: function () 
#>     initialize: function (path, access = "rw") 
#>     list_content: function (...) 
#>     read: function (name, ...) 
#>     remove: function (file) 
#>     write: function (x, file, ...) 
#>   Private:
#>     path: /home/runner/work/_temp/Library/connector/demo_trial/adam
class(db_subclass)
#> [1] "subclass"     "Connector_fs" "R6"