Connector for file storage

The connector_fs class is a file storage connector for accessing and manipulating files any file storage solution. The default implementation includes methods for files stored on local or network drives.

Super class

connector::connector -> connector_fs

Active bindings

path

character Path to the file storage

Methods

Public methods

connector_fs$new()

connector_fs$cnt_download()

connector_fs$cnt_upload()

connector_fs$cnt_create_directory()

connector_fs$cnt_remove_directory()

connector_fs$clone()

Inherited methods

connector::connector$cnt_list_content()

connector::connector$cnt_read()

connector::connector$cnt_remove()

connector::connector$cnt_write()

connector::connector$print()

Method new()

Initializes the connector for file storage.

Usage

connector_fs$new(path, access = "rw", extra_class = NULL)

Arguments

path

character Path to the file storage

access

character Access type ("rw" by default). Checked using checkmate::assert_directory_exists.

extra_class

character Extra class to assign to the new connector.

Method cnt_download()

Download content from the file storage. See also cnt_download.

Usage

connector_fs$cnt_download(name, file = basename(name), ...)

Arguments

name

character Name of the content to read, write, or remove. Typically the table name.

file

character Path to the file to download to or upload from

...

Additional arguments passed to the method for the individual connector.

Returns

invisible file

Method cnt_upload()

Upload a file to the file storage. See also cnt_upload.

Usage

connector_fs$cnt_upload(file, name = basename(file), ...)

Arguments

file

character Path to the file to download to or upload from

name

character Name of the content to read, write, or remove. Typically the table name.

...

Additional arguments passed to the method for the individual connector.

Returns

invisible self.

Method cnt_create_directory()

Create a directory in the file storage. See also cnt_create_directory.

Usage

connector_fs$cnt_create_directory(name, ...)

Arguments

name

character The name of the directory to create

...

Additional arguments passed to the method for the individual connector.

Returns

invisible self.

Method cnt_remove_directory()

Remove a directory from the file storage. See also cnt_remove_directory.

Usage

connector_fs$cnt_remove_directory(name, ...)

Arguments

name

character The name of the directory to remove

...

Additional arguments passed to the method for the individual connector.

Returns

invisible self.

Method clone()

The objects of this class are cloneable with this method.

Usage

connector_fs$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Create file storage connector

cnt <- connector_fs$new(tempdir())

cnt
#> <connector_fs>
#> Inherits from: <connector>
#> Registered methods:
#> • `cnt_create_directory.connector_fs()`
#> • `cnt_download.connector_fs()`
#> • `cnt_list_content.connector_fs()`
#> • `cnt_read.connector_fs()`
#> • `cnt_remove.connector_fs()`
#> • `cnt_remove_directory.connector_fs()`
#> • `cnt_upload.connector_fs()`
#> • `cnt_write.connector_fs()`
#> Specifications:
#> • path: /tmp/RtmpXHHJ6l

# List content

cnt$cnt_list_content()
#> [1] "bslib-84fe5b9ef621e42e9aded10b166b6385"
#> [2] "downlit"                               
#> [3] "example.txt"                           
#> [4] "file167734f466f7"                      
#> [5] "file16775eb97de6"                      
#> [6] "file16775f855bc4"                      
#> [7] "iris.csv"                              
#> [8] "iris.rds"                              

# Write to the connector

cnt$cnt_write(iris, "iris.rds")

# Check it is there

cnt$cnt_list_content()
#> [1] "bslib-84fe5b9ef621e42e9aded10b166b6385"
#> [2] "downlit"                               
#> [3] "example.txt"                           
#> [4] "file167734f466f7"                      
#> [5] "file16775eb97de6"                      
#> [6] "file16775f855bc4"                      
#> [7] "iris.csv"                              
#> [8] "iris.rds"                              

# Read the result back

cnt$cnt_read("iris.rds") |>
  head()
#>   Sepal.Length Sepal.Width Petal.Length Petal.Width Species
#> 1          5.1         3.5          1.4         0.2  setosa
#> 2          4.9         3.0          1.4         0.2  setosa
#> 3          4.7         3.2          1.3         0.2  setosa
#> 4          4.6         3.1          1.5         0.2  setosa
#> 5          5.0         3.6          1.4         0.2  setosa
#> 6          5.4         3.9          1.7         0.4  setosa