List available content from the connector

Generic implementing of how to list all content available for different connectors:

ConnectorDBI: Uses DBI::dbListTables() to list the tables in a DBI connection.

ConnectorFS: Uses list.files() to list all files at the path of the connector.

ConnectorLogger: Logs the list operation and calls the underlying connector method.

Usage

list_content_cnt(connector_object, ...)

# S3 method for class 'ConnectorDBI'
list_content_cnt(connector_object, ...)

# S3 method for class 'ConnectorFS'
list_content_cnt(connector_object, ...)

# S3 method for class 'ConnectorLogger'
list_content_cnt(connector_object, ...)

Arguments

connector_object

Connector The connector object to use.

...

Additional arguments passed to the method for the individual connector.

Value

A character vector of content names

Examples

# List tables in a DBI database
cnt <- connector_dbi(RSQLite::SQLite())

cnt |>
  list_content_cnt()
#> character(0)

# List content in a file storage
cnt <- connector_fs(tempdir())

cnt |>
  list_content_cnt()
#>  [1] "bslib-2ce9eb5d8d7866b4f2d82a9ef19eb08f"                                                             
#>  [2] "downlit"                                                                                            
#>  [3] "file1e9c1f52fc30.yml"                                                                               
#>  [4] "file1e9c21bf40cc.yml"                                                                               
#>  [5] "file1e9c2ac0d895"                                                                                   
#>  [6] "file1e9c3ae9cace"                                                                                   
#>  [7] "file1e9c44ab21de"                                                                                   
#>  [8] "file1e9c47f7093d"                                                                                   
#>  [9] "file1e9c618451d5"                                                                                   
#> [10] "file1e9c700f79a3"                                                                                   
#> [11] "file1e9c77f530be"                                                                                   
#> [12] "file1e9c78350737"                                                                                   
#> [13] "file1e9cb265e10"                                                                                    
#> [14] "repos_https%3A%2F%2Fpackagemanager.posit.co%2Fcran%2F__linux__%2Fnoble%2Flatest%2Fsrc%2Fcontrib.rds"

# Only list CSV files using the pattern argument of list.files

cnt |>
  list_content_cnt(pattern = "\\.csv$")
#> character(0)

# Add logging to a connector and list contents
folder <- withr::local_tempdir()
cnt <- connectors(data = connector_fs(folder)) |> add_logs()
#> Error in validate_resource(self): Invalid file system connector: /tmp/RtmpXfFLTE/file1e9c7e0b58a does not
#> exist.

cnt$data |>
  write_cnt(iris, "iris.csv")
#> Error in method_error_msg(connector_object): Method not implemented for class <NULL>
#> ℹ See the customize (`vignette(connector::customize)`) vignette on how to
#>   create custom connectors and methods

cnt$data |>
  list_content_cnt()
#> Error in method_error_msg(connector_object): Method not implemented for class <NULL>
#> ℹ See the customize (`vignette(connector::customize)`) vignette on how to
#>   create custom connectors and methods