List available content from the connector

Generic implementing of how to list all content available for different connectors:

connector_dbi: Uses DBI::dbListTables() to list the tables in a DBI connection.

connector_fs: Uses list.files() to list all files at the path of the connector.

Usage

list_content_cnt(connector_object, ...)

# S3 method for class 'connector_dbi'
list_content_cnt(connector_object, ...)

# S3 method for class 'connector_fs'
list_content_cnt(connector_object, ...)

Arguments

connector_object

connector The connector object to use.

...

Additional arguments passed to the method for the individual connector.

Value

A character vector of content names

Examples

# List tables in a DBI database
cnt <- connector_dbi$new(RSQLite::SQLite())

cnt |>
  list_content_cnt()
#> character(0)

# List content in a file storage
cnt <- connector_fs$new(tempdir())

cnt |>
  list_content_cnt()
#> [1] "bslib-9f68e11c24c13beda8ab56edec83b1cf"
#> [2] "downlit"                               
#> [3] "file1b6b13c8359a"                      
#> [4] "file1b6b24652ad8"                      
#> [5] "file1b6b34ea4ec4"                      
#> [6] "file1b6b376f6579"                      
#> [7] "iris.rds"                              

# Only list CSV files using the pattern argument of list.files

cnt |>
  list_content_cnt(pattern = "\\.csv$")
#> character(0)